Safer Pop/Up by limiting capacity
Pop/Up now limit the capacity of the array to its len to force
allocating a new array when growing.

I'm not sure this is the right thing to do as it may provoc more
allocation when traversing a document with Property/Index/Pop/Up.
In fact I should add a test to see what really happens...